refacto: split views file
