+++ gsl_notification/urls.py
@@ -1,15 +1,17 @@
 from django.urls import path
 
-from gsl_notification.views.views import (
+from gsl_notification.views.modele_arrete_views import (
     CreateModelArreteWizard,
     DuplicateModeleArete,
     ModeleArreteListView,
     UpdateModeleArrete,
+    delete_modele_arrete_view,
+)
+from gsl_notification.views.views import (
     change_arrete_view,
     create_arrete_signe_view,
     delete_arrete_signe_view,
     delete_arrete_view,
-    delete_modele_arrete_view,
     documents_view,
     download_arrete,
     download_arrete_signe,
